+++ main_test.go
@@ -23,6 +23,52 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
+// BenchmarkMain is the benchmarking entrypoint.
+func BenchmarkMain(b *testing.B) {
+	tx, txErr := DefaultDb().Begin()
+	if txErr != nil {
+		b.Error("Unable to create transaction")
+		b.FailNow()
+	}
+	if tx == nil {
+		b.Error("`tx` is nil")
+		b.FailNow()
+	}
+
+	defer func() {
+		if tx != nil {
+			if rollbackErr := tx.Rollback(); rollbackErr != nil {
+				b.Errorf("Error rolling back transaction: %v", rollbackErr)
+				b.FailNow()
+			}
+		}
+	}()
+
+	seedErr := seedObjects(5000, tx)
+	if seedErr != nil {
+		b.Errorf("Error seeding objects: %v", seedErr)
+		b.FailNow()
+	}
+
+	manualBefore := time.Now()
+	_, manualErr := readManual(tx)
+	manualAfter := time.Now()
+	if manualErr != nil {
+		b.Errorf("Error using manual query: %v", manualErr)
+		b.FailNow()
+	}
+
+	ormBefore := time.Now()
+	_, ormErr := readOrm(tx)
+	ormAfter := time.Now()
+	if ormErr != nil {
+		b.Errorf("Error using orm: %v", ormErr)
+		b.FailNow()
+	}
+
+	b.Logf("Benchmark Test Results: Manual: %v vs. Orm: %v\n", manualAfter.Sub(manualBefore), ormAfter.Sub(ormBefore))
+}
+
 func dbConnectionFromEnvironment() *DbConnection {
 	dbHost := os.Getenv("DB_HOST")
 	dbSchema := os.Getenv("DB_SCHEMA")
